Add a notification when the time limit is already up. Also, add the code for verify,unverify, and exitgroup command. Furthermore, add the date in the history command.

+++ src/main/java/bot/command/RemoveAllCommand.java
@@ -14,6 +14,11 @@ public class RemoveAllCommand implements Command {
 
     @Override
     public String execute() {
+
+        if (co.getExceedTimeLimitStatus()) {
+            return "Time's already up! You can't add items anymore!";
+        }
+
         try {
             co.deleteAllOrderFromUser(telegram_id);
             return "Successfully removing all the orders!";
